Refactor to consolidate storage modules
