+++ src/App.tsx
@@ -6,8 +6,12 @@ import { Sun, Moon } from 'lucide-solid';
 import ScoreboardOCR from '#c/ScoreboardOCR';
 import RegionProfileManager from '#c/RegionProfileManager';
 import GameRecordsTable from '#c/GameRecordsTable';
-import { triggerUploadDialog, handleFileUpload } from '#utils/gameStorage';
-import { loadSettings, saveSettings } from '#utils/storage';
+import {
+	triggerUploadDialog,
+	handleFileUpload,
+	loadSettings,
+	saveSettings,
+} from '#utils/storage';
 import '#styles/App';
 
 type ViewMode = 'ocr' | 'records' | 'regions';
